+++ packages/api/src/services/auth.service.ts
@@ -1,8 +1,13 @@
 import { TRPCError } from '@trpc/server'
 
 import type { Database } from '@yukinu/db'
-import type { AuthValidator } from '@yukinu/validators/auth'
-import { invalidateSessionTokens, Password } from '@yukinu/auth'
+import type { AuthModels } from '@yukinu/validators/auth'
+import {
+  createSessionCookie,
+  invalidateSession,
+  invalidateSessionTokens,
+  Password,
+} from '@yukinu/auth'
 
 import type {
   IAccountRepository,
@@ -14,96 +19,127 @@ import type {
 export class AuthService implements IAuthService {
   private readonly _password: Password
 
-  constructor(
+  public constructor(
     private readonly _db: Database,
-    private readonly _accountRepo: IAccountRepository,
-    private readonly _profileRepo: IProfileRepository,
-    private readonly _userRepo: IUserRepository,
+    private readonly _account: IAccountRepository,
+    private readonly _profile: IProfileRepository,
+    private readonly _user: IUserRepository,
   ) {
     this._password = new Password()
   }
 
-  async register(
-    data: AuthValidator.RegisterBody,
-  ): Promise<{ id: IUserRepository.UserType['id'] }> {
-    const { username, email } = data
+  public async login(
+    input: AuthModels.LoginInput,
+    headers: Headers,
+    resHeaders: Headers,
+  ): Promise<AuthModels.LoginOutput> {
+    const { identifier, password } = input
 
-    const [user] = await this._userRepo.findBy([{ username }, { email }], {}, 1)
-    if (user)
+    const [existingUser] = await this._user.findBy(
+      [{ email: identifier }, { username: identifier }],
+      {},
+      1,
+    )
+    if (!existingUser)
       throw new TRPCError({
-        code: 'CONFLICT',
-        message: 'Username or email already exists',
+        code: 'UNAUTHORIZED',
+        message: 'Invalid credentials',
       })
 
+    const [existingAccount] = await this._account.findBy(
+      [{ userId: existingUser.id, provider: 'credentials' }],
+      {},
+      1,
+    )
+    if (
+      !existingAccount?.password ||
+      !(await this._password.verify(existingAccount.password, password))
+    )
+      throw new TRPCError({
+        code: 'UNAUTHORIZED',
+        message: 'Invalid credentials',
+      })
+
+    const { token, expires, cookie } = await createSessionCookie(
+      existingUser.id,
+      headers,
+    )
+
+    resHeaders.append('Set-Cookie', cookie)
+    return { token, expires }
+  }
+
+  public async logout(headers: Headers, resHeaders: Headers): Promise<void> {
+    const cookie = await invalidateSession(headers)
+    if (cookie) resHeaders.append('Set-Cookie', cookie)
+  }
+
+  public async register(
+    input: AuthModels.RegisterInput,
+  ): Promise<AuthModels.RegisterOutput> {
+    const { email, username } = input
+
+    const [existingUser] = await this._user.findBy(
+      [{ email }, { username }],
+      {},
+      1,
+    )
+    if (existingUser)
+      throw new TRPCError({
+        code: 'CONFLICT',
+        message: 'Email or username already in use',
+      })
+
+    const password = await this._password.hash(input.password)
+
     return this._db.transaction(async (tx) => {
-      const user = await this._userRepo.create({ username, email }, tx)
-      if (!user) return tx.rollback()
+      const { id: userId } = await this._user.create({ email, username }, tx)
 
-      const provider = 'credentials'
-      const password = await this._password.hash(data.password)
-
-      await this._accountRepo.create(
-        { userId: user.id, provider, accountId: user.id, password },
+      await this._account.create(
+        { userId, provider: 'credentials', accountId: userId, password },
         tx,
       )
 
-      await this._profileRepo.create(
-        { id: user.id, fullName: data.username },
-        tx,
-      )
+      await this._profile.create({ id: userId, fullName: username }, tx)
 
-      return { id: user.id }
+      return { userId }
     })
   }
 
-  async changePassword(
-    userId: string,
-    data: AuthValidator.ChangePasswordBody,
-  ): Promise<{ id: IUserRepository.UserType['id'] }> {
-    const { currentPassword, newPassword, isLogOutOtherSessions } = data
+  public async changePassword(
+    input: AuthModels.ChangePasswordInput,
+  ): Promise<AuthModels.ChangePasswordOutput> {
+    const { userId, currentPassword, newPassword, isLogOutOtherSessions } =
+      input
 
-    const [existedAccount] = await this._accountRepo.findBy(
-      [{ accountId: userId, provider: 'credentials' }],
+    const [existingAccount] = await this._account.findBy(
+      [{ userId, provider: 'credentials' }],
       {},
       1,
     )
 
-    const provider = 'credentials'
-    const newPasswordHash = await this._password.hash(newPassword)
+    if (
+      existingAccount?.password &&
+      !(await this._password.verify(
+        existingAccount.password,
+        currentPassword ?? '',
+      ))
+    )
+      throw new TRPCError({
+        code: 'UNAUTHORIZED',
+        message: 'Invalid current password',
+      })
 
-    const result = await this._db.transaction(async (tx) => {
-      if (!existedAccount) {
-        await this._accountRepo.create(
-          { userId, provider, accountId: userId, password: newPasswordHash },
-          tx,
-        )
-        return { id: userId }
-      }
-
-      const { id, password } = existedAccount
-      if (password === null || !currentPassword)
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'Current password is required',
-        })
-
-      if (!(await this._password.verify(password, currentPassword)))
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'Current password is incorrect',
-        })
-
-      if (currentPassword === newPassword)
-        throw new TRPCError({
-          code: 'BAD_REQUEST',
-          message: 'New password must be different from the current password',
-        })
-
-      await this._accountRepo.update(id, { password: newPasswordHash }, tx)
-      return { id: userId }
+    const password = await this._password.hash(newPassword)
+    await this._account.create({
+      userId,
+      provider: 'credentials',
+      accountId: userId,
+      password,
     })
 
     if (isLogOutOtherSessions) await invalidateSessionTokens(userId)
-    return result
+
+    return { userId }
   }
 }