+++ packages/auth/src/react.tsx
@@ -3,25 +3,12 @@
 import * as React from 'react'
 import { useQuery } from '@tanstack/react-query'
 
-import type { Providers } from '@/config'
 import type { Session, User } from '@/types'
 
-type AuthProviders =
-  | 'credentials'
-  | (Providers extends never ? undefined : Providers)
 type SessionResult = Omit<Session, 'token' | 'userAgent' | 'ipAddress'>
 
 type SessionContextValue = {
-  signIn: <
-    TProvider extends AuthProviders,
-    TData extends { token: string; expires: string },
-  >(
-    provider: TProvider,
-    ...args: TProvider extends 'credentials'
-      ? [{ identifier: string; password: string }]
-      : [{ redirectUrl?: string }?]
-  ) => Promise<TData | null>
-  signOut: (opts?: { redirectUrl: string }) => Promise<void>
+  refresh: () => Promise<unknown>
 } & (
   | { status: 'loading'; session: SessionResult }
   | { status: 'unauthenticated'; session: SessionResult & { user: null } }
@@ -48,7 +35,7 @@ function SessionProvider({
 }>) {
   const {
     data: session,
-    isLoading,
+    status,
     refetch,
   } = useQuery({
     queryKey: ['auth', 'session'],
@@ -58,68 +45,25 @@ function SessionProvider({
         headers: { 'Content-Type': 'application/json' },
         signal,
       })
-      if (!res.ok) return { user: null, expires: new Date() }
 
-      return (await res.json()) as SessionResult
+      const session = (await res.json()) as SessionResult
+      if (!session.user) throw new Error('No active session')
+      return session
     },
     initialData: _session ?? undefined,
-    staleTime: 1000 * 60 * 5, // 5 minutes
+    staleTime: 1000 * 60 * 10, // 10 minutes
+    retry: false,
   })
 
-  const status = React.useMemo(() => {
-    if (isLoading) return 'loading'
-    return session?.user ? 'authenticated' : 'unauthenticated'
-  }, [isLoading, session])
+  const value = React.useMemo(() => {
+    const statusMap = {
+      pending: 'loading',
+      error: 'unauthenticated',
+      success: 'authenticated',
+    } as const
 
-  const signIn = React.useCallback(
-    async <
-      TProvider extends AuthProviders,
-      TData extends { token: string; expires: string },
-    >(
-      provider: TProvider,
-      ...args: TProvider extends 'credentials'
-        ? [{ identifier: string; password: string }]
-        : [{ redirectUrl?: string }?]
-    ): Promise<TData | null> => {
-      if (provider === 'credentials') {
-        const res = await fetch(`${base}/api/auth/sign-in`, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(args[0]),
-        })
-
-        if (!res.ok) {
-          const text = await res.text()
-          throw new Error(text)
-        } else {
-          await refetch()
-          const json = (await res.json()) as TData
-          return json
-        }
-      } else {
-        const redirectUrl = (args[0] as { redirectUrl?: string }).redirectUrl
-        window.location.href = `${base}/api/auth/${provider}${
-          redirectUrl ? `?redirectUrl=${encodeURIComponent(redirectUrl)}` : ''
-        }`
-        return null
-      }
-    },
-    [base, refetch],
-  )
-
-  const signOut = React.useCallback(
-    async (opts?: { redirectUrl: string }) => {
-      await fetch(`${base}/api/auth/sign-out`, { method: 'POST' })
-      await refetch()
-      if (opts?.redirectUrl) window.location.href = opts.redirectUrl
-    },
-    [base, refetch],
-  )
-
-  const value = React.useMemo(
-    () => ({ status, session, signIn, signOut }),
-    [status, session, signIn, signOut],
-  ) as SessionContextValue
+    return { status: statusMap[status], session, refresh: refetch }
+  }, [status, session, refetch]) as SessionContextValue
 
   return <SessionContext value={value}>{children}</SessionContext>
 }