+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @yukinu/dashboard
 
+## 0.2.4
+
+### Patch Changes
+
+- [#169](https://github.com/tiesen243/yukinu/pull/169) [`085ce77`](https://github.com/tiesen243/yukinu/commit/085ce7734c7bf53c0e64e3c0418399d1365bd309) Thanks [@tiesen243](https://github.com/tiesen243)! - update ui
+
+- Updated dependencies [[`b4893aa`](https://github.com/tiesen243/yukinu/commit/b4893aa564c30e295ad2f443c20f295f2bcd84e8), [`9dc7549`](https://github.com/tiesen243/yukinu/commit/9dc754977d85fe399e7a271002e8cce393ba97f7), [`085ce77`](https://github.com/tiesen243/yukinu/commit/085ce7734c7bf53c0e64e3c0418399d1365bd309)]:
+  - @yukinu/validators@0.1.4
+  - @yukinu/api@0.1.5
+  - @yukinu/auth@0.1.4
+  - @yukinu/db@0.1.4
+
+## 0.2.3
+
+### Patch Changes
+
+- [#167](https://github.com/tiesen243/yukinu/pull/167) [`3525430`](https://github.com/tiesen243/yukinu/commit/352543000a4e985d0302b395205ad9b06f5246df) Thanks [@tiesen243](https://github.com/tiesen243)! - decrease dashboard image size
+
+- [#167](https://github.com/tiesen243/yukinu/pull/167) [`f5a8716`](https://github.com/tiesen243/yukinu/commit/f5a871614db21c11f2e51752e0e8de6f94a85fb2) Thanks [@tiesen243](https://github.com/tiesen243)! - optimize image
+
 ## 0.2.2
 
 ### Patch Changes